+++ src/main/java/com/love/mapper/user/UserVideoMapper.java
@@ -15,7 +15,7 @@ public interface UserVideoMapper {
 
     // 分页查询用户视频
     @Select("select * from video where user_id=#{userId} order by update_time desc")
-    List<Video> list(Long userId, Integer categoryId, String state);
+    List<Video> list(Long userId, Long categoryId, String state);
 
     // 修改用户视频信息
     void update(Video userVideo);
