+++ src/main/java/com/love/service/VideoService.java
@@ -6,7 +6,7 @@ import java.util.List;
 
 public interface VideoService {
     // 查询所有视频
-    List<Video> list(Integer categoryId);
+    List<Video> list(Long categoryId);
     // 根据id查询视频
     Video findById(Integer id);
     //用户收藏列表
